backend/requirements/base.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0
python-decouple>=3.8
psycopg2-binary>=2.9
dj-database-url>=2.0
anthropic>=0.40.0
